requirements.txt
pytest~=7.4.2
selene>=2.0.0b4
allure-pytest==2.9.45
requests==2.31.0
jsonschema==4.20.0